+++ docs/CHANGELOG.md
@@ -14,6 +14,7 @@ Version 1.6.1 (not yet released)
 - Improve crash reporting for Dash Editor and launcher
 - Do not print redundant new lines in `-win32-console` server mode and optimize it a little
 - Fix unwanted server-side weapon fire when switching from a continous weapon or a weapon with burst fire
+- Prevent shooting and reloading while switching weapon server-side
 
 Version 1.6.0 (released 2021-01-07)
 --------------------------------